renaming ActionProxy to APIProxy
