+++ lib/ShopifyAPI.js
@@ -1,9 +1,9 @@
 "use strict";
 
-import API from "./utilities/ActionProxy";
+import APIProxy from "./utilities/APIProxy";
 
 module.exports = {
-  Product: new API(require("./actions/ProductActions"), require("./stores/ProductsStore")),
-  Session: new API(require("./actions/SessionActions"), null),
-  Shop: new API(require("./actions/ShopActions"), require("./stores/ShopStore"))
+  Product:  new APIProxy(require("./actions/ProductActions"), require("./stores/ProductsStore")),
+  Session:  new APIProxy(require("./actions/SessionActions"), null),
+  Shop:     new APIProxy(require("./actions/ShopActions"), require("./stores/ShopStore"))
 };
