+++ lib/utilities/__tests__/APIProxyTest.js
@@ -1,29 +1,28 @@
 "use strict";
 
-jest.dontMock("../ActionProxy");
+jest.dontMock("../APIProxy");
 jest.dontMock("../BaseActions");
 jest.dontMock("../../actions/ProductActions");
 
-describe("ActionProxy", () => {
+describe("APIProxy", () => {
   var api, actions, store;
 
   beforeEach(() => {
-    const ActionProxy = require("../ActionProxy");
+    const APIProxy = require("../APIProxy");
 
     actions       = require("../../actions/ProductActions");
     actions.fetch = jest.genMockFunction();
 
     store = require("../../stores/ProductsStore");
-    api   = new ActionProxy(actions, store);
+    api   = new APIProxy(actions, store);
   });
 
-  it("exposes action methods to the API", () => {
-    ["clear", "fetch", "fetchByCollection", "fetchAllSince"].forEach(method => {
-      expect(api[method]).toBeDefined(`${method} not defined`);
-    });
+  it("decorates the proxy object with action methods", () => {
+    let actionMethods = actions.constructor.actionMethods;
+    actionMethods.forEach(method => expect(api[method]).toBeDefined(`${method} not defined`));
   });
 
-  it("delegates action method calls to the action object", () => {
+  it("proxies action method calls to the action object", () => {
     api.fetch();
     expect(actions.fetch).toBeCalled();
   });
